% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDAM}
\format{
## 'fire'
A data frame with 2 columns and 16 rows
\describe{
  \item{DISTANCE}{distance}
  \item{DAMAGE}{damage}
  ...
}
}
\source{
MATH4753
}
\usage{
fire
}
\description{
FIREDAM
}
\keyword{datasets}
