% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{MyBin}
\usage{
mybin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{vector}

\item{p}{probability}
}
\value{
This function returns a barplot
}
\description{
Function developed by SNT for Lab 5 MATH 4753
}
\examples{
mybin(iter=10000,n=10, p=0.7)
}
