% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot2}
\alias{myboot2}
\title{MyBoot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iterations}

\item{x}{sample}

\item{fun}{function name "mean, sd, var, IQR" etc}

\item{alpha}{confidence interval}

\item{cx}{text size}

\item{...}{other graphical descriptions}
}
\value{
This function returns a bargraph showing bootstrap intervals as well as list of statistics
}
\description{
Function develpoed by SNT for Lab 9 MATH 4753
}
\examples{
\dontrun{myboot2(x=ddt)}

}
