% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{MyCI}
\usage{
myci(x, a = 0.95, n)
}
\arguments{
\item{x}{= sample}

\item{a}{= conf interval, as a decimal}

\item{n}{= sample size}
}
\value{
range of values
}
\description{
Function developed by SNT for Lab 11 MATH 4753
}
\examples{
\dontrun{myci(x,n=30)}
}
