% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Uniform Distribution CLT}
\usage{
mycltu(n, iter, a = 0, b = 10, x)
}
\arguments{
\item{n}{number of samples taken}

\item{iter}{number per sample}

\item{a}{min}

\item{b}{max}

\item{x}{normal curve sample}
}
\value{
This function returns a histogram with density curves and normal distribution density
}
\description{
Function developed by SNT for Lab 8 MATH 4753
}
\examples{
mycltu(10,1000)
mycltu(10,1000,0,10)
}
