% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{MyMaxLik}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{logfunction}

\item{x}{1}

\item{param}{2}

\item{...}{other graphical descriptors}
}
\value{
graph
}
\description{
This function was developed by SNT for Lab 10 MATH 4753
}
\examples{
\dontrun{mymaxlik(lfun=logbin,x=c(9,9,1,9,9,9)}
}
