% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Lab 6 myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{- mean of normal distribution}

\item{sigma}{- sd of normal distribution}

\item{a}{- Y ~ N(Y less than a)}
}
\value{
a plot showing probability of Y ~ N(Y less than a), and calculated probability
}
\description{
Lab 6 myncurve
}
\examples{
myncurve(1,1,2)
}
