% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{MyPValue}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05, x)
}
\arguments{
\item{t0}{t(not)}

\item{xmax}{xmax}

\item{n}{n value}

\item{alpha}{confidence interval}

\item{x}{vector of quantiles (dt function)}
}
\value{
This function returns a q and pvalue as well as a density plot with areas outside the confidence interval highlighted.
}
\description{
Function created by SNT for Lab 12 MATH 4753
}
\examples{
\dontrun{mypvalue(tcalc,n=30,alpha=0.05)}
}
