% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Project 1: Plane Sales}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{= total number of seat on plane}

\item{gamma}{= acceptable probbility of overselling}

\item{p}{= probability passenger arrives}
}
\value{
This function returns the listed parameters: nd, nc, N, gamma, p and generates a discrete binomial plot, continuous normal plot
}
\description{
Function created by SNT for Project 1 MATH 4753
}
\examples{
ntickets(400,.02,.95)
}
