% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{MyNCurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{- mean of normal distribution}

\item{sigma}{- sd of normal distribution}

\item{a}{- Y ~ N(Y less than a)}
}
\value{
This function returns a plot showing probability of Y ~ N(Y less than a), and calculated probability
}
\description{
Function developed by SNT for Lab 6 MATH 4753
}
\examples{
\dontrun{myncurve(1,1,2)}
}
